test4

+++ utils.c
@@ -55,7 +55,7 @@ int is_digit(char c)
  * @size: the number indicating the type to be casted to convert.
  * Return: the casted value of numero
  */
-long int convert_size_number(long int num, int size)
+long int convert_size_number(int num, int size)
 {
 	if (size == S_LONG)
 		return (num);
@@ -71,12 +71,12 @@ long int convert_size_number(long int num, int size)
  * @size: the number indicating the type to be casted.
  * Return: the casted value of numero.
  */
-long int convert_size_unsgnd(unsigned long int num, int size)
+int convert_size_unsgnd(int num, int size)
 {
 	if (size == S_LONG)
 		return (num);
 	else if (size == S_SHORT)
 		return ((unsigned short)num);
 
-	return ((unsigned int)num);
+	return ((int)num);
 }
